Upgraded to latest gradle and plugins
Resolved some complications issues.

The kotlin plugin in intellij is not compatiable with gradle plugin
projects using the kotlin-dsl gradle plugin.
