+++ gradle.properties
@@ -1,3 +1,4 @@
+# suppress inspection "UnusedProperty" for whole file
 org.gradle.parallel=true
 org.gradle.caching=true
 version=0.0.6-SNAPSHOT
